refactor(web): normalize component folder naming convention
