+++ apps/web/app/[locale]/dashboard/app-url/[teamId]/page.tsx
@@ -7,17 +7,17 @@ import { useOrganizationTeams } from '@app/hooks/features/useOrganizationTeams';
 import { useAtomValue } from 'jotai';
 import { useTranslations } from 'next-intl';
 import { useParams } from 'next/navigation';
-import React, { useMemo } from 'react';
+import { useMemo } from 'react';
 import { ArrowLeftIcon } from '@radix-ui/react-icons';
 import { useRouter } from 'next/navigation';
 import { Breadcrumb, Container } from '@/lib/components';
 import { DashboardHeader } from '../../team-dashboard/[teamId]/components/dashboard-header';
 import { GroupByType, useReportActivity } from '@/app/hooks/features/useReportActivity';
 import { Card } from '@components/ui/card';
-import { ProductivityHeader } from '../components/ProductivityHeader';
-import { ProductivityChart } from '../components/ProductivityChart';
-import { ProductivityStats } from '../components/ProductivityStats';
-import { ProductivityTable } from '../components/ProductivityTable';
+import { ProductivityHeader } from '../components/productivity-header';
+import { ProductivityChart } from '../components/productivity-chart';
+import { ProductivityStats } from '../components/productivity-stats';
+import { ProductivityTable } from '../components/productivity-table';
 import { useLocalStorageState, useModal } from '@/app/hooks';
 import { ProductivityApplicationTable, ProductivityEmployeeTable, ProductivityProjectTable } from '../components';
