+++ apps/web/app/[locale]/dashboard/app-url/components/index.tsx
@@ -1,4 +1,4 @@
-export * from './ProductivityTable'
-export * from './productivity-application/ProductivityApplicationTable'
-export * from './productivity-employee/ProductivityEmployeeTable'
-export * from './productivity-project'
+export * from './productivity-table';
+export * from './productivity-application/productivity-application-table';
+export * from './productivity-employee/productivity-employee-table';
+export * from './productivity-project';
