+++ apps/web/app/[locale]/profile/[memberId]/components/user-profile-detail.tsx
