+++ apps/web/app/[locale]/time-and-activity/Components/activity-level-cell.tsx
