+++ apps/web/app/[locale]/time-and-activity/Components/activity-table-skeleton.tsx
