+++ apps/web/app/[locale]/time-and-activity/Components/earnings-cell.tsx
