+++ apps/web/app/[locale]/time-and-activity/Components/empty-time-activity.tsx
