+++ apps/web/app/[locale]/time-and-activity/Components/index.tsx
@@ -1,6 +1,6 @@
 export * from './page-component';
 export * from './time-activity-header';
 export * from './date-range-picker-time-activity';
-export * from './GroupBySelectTimeActivity';
-export * from './ViewSelect'
-export * from './time-activityFilterPopover';
+export * from './group-by-select-time-activity';
+export * from './view-select';
+export * from './time-activity-filter-popover';
