+++ apps/web/app/[locale]/time-and-activity/Components/time-activity-table-skeleton.tsx
