+++ apps/web/app/[locale]/timesheet/[memberId]/components/calendar-view.tsx
@@ -3,13 +3,13 @@ import { statusColor } from '@/lib/components';
 import { DisplayTimeForTimesheet, TaskNameInfoDisplay, TotalDurationByDate, TotalTimeDisplay } from '@/lib/features';
 import { Accordion, AccordionContent, AccordionItem, AccordionTrigger } from '@components/ui/accordion';
 import { TranslationHooks, useTranslations } from 'next-intl';
-import { EmployeeAvatar, ProjectLogo } from './CompactTimesheetComponent';
+import { EmployeeAvatar, ProjectLogo } from './compact-timesheet-component';
 import { formatDate } from '@/app/helpers';
 import { ClockIcon, CodeSquareIcon } from 'lucide-react';
 import { cn } from '@/lib/utils';
-import MonthlyTimesheetCalendar from './MonthlyTimesheetCalendar';
+import MonthlyTimesheetCalendar from './monthly-timesheet-calendar';
 import { useTimelogFilterOptions } from '@/app/hooks';
-import WeeklyTimesheetCalendar from './WeeklyTimesheetCalendar';
+import WeeklyTimesheetCalendar from './weekly-timesheet-calendar';
 import { TimesheetLog } from '@/app/interfaces';
 import TimesheetSkeleton from '@components/shared/skeleton/TimesheetSkeleton';
 import { Checkbox } from '@components/ui/checkbox';
