+++ apps/web/app/[locale]/timesheet/[memberId]/components/compact-timesheet-component.tsx
