+++ apps/web/app/[locale]/timesheet/[memberId]/components/selection-bar.tsx
