+++ apps/web/app/[locale]/timesheet/[memberId]/components/timesheet-detail-modal.tsx
@@ -1,7 +1,7 @@
 import { TimesheetLog, TimesheetStatus } from '@/app/interfaces';
 import { Modal, statusColor } from '@/lib/components';
 import { AnimatedEmptyState } from '@components/ui/empty-state';
-import { TimesheetCardDetail } from './TimesheetCard';
+import { TimesheetCardDetail } from './timesheet-card';
 import { TranslationHooks, useTranslations } from 'next-intl';
 import { TimesheetDetailMode } from '../page';
 import { Accordion, AccordionContent, AccordionItem, AccordionTrigger } from '@components/ui/accordion';
@@ -9,7 +9,7 @@ import { cn } from '@/lib/utils';
 import { useTimesheet } from '@/app/hooks/features/useTimesheet';
 import { Badge } from '@components/ui/badge';
 import { TaskNameInfoDisplay, TotalTimeDisplay } from '@/lib/features';
-import { EmployeeAvatar, ProjectLogo } from './CompactTimesheetComponent';
+import { EmployeeAvatar, ProjectLogo } from './compact-timesheet-component';
 import { groupBy } from '@/app/helpers/array-data';
 
 export interface IAddTaskModalProps {
