+++ apps/web/app/[locale]/timesheet/[memberId]/page.tsx
@@ -37,9 +37,9 @@ import type { IconBaseProps } from 'react-icons';
 import { differenceBetweenHours, getGreeting, secondsToTime } from '@/app/helpers';
 import { useTimesheet } from '@/app/hooks/features/useTimesheet';
 import { endOfMonth, startOfMonth } from 'date-fns';
-import TimesheetDetailModal from './components/TimesheetDetailModal';
+import TimesheetDetailModal from './components/timesheet-detail-modal';
 import { useTimesheetPagination } from '@/app/hooks/features/useTimesheetPagination';
-import TimesheetPagination from './components/TimesheetPagination';
+import TimesheetPagination from './components/timesheet-pagination';
 import { useTimesheetFilters } from '@/app/hooks/features/useTimesheetFilters';
 import { useTimesheetViewData } from '@/app/hooks/features/useTimesheetViewData';
 import { IconsSearch } from '@/icons';
